requirements.txt
aiogram==3.20.0.post0
aiohttp==3.11.11
python-dotenv==1.1.0
redis==6.2.0
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
spacy==3.7.5
